Move types from `schema` to `types` package.
